chore(queue): move syncMatriculas to core
