+++ apps/core/src/queue/setup.ts
@@ -4,6 +4,7 @@ import { createWorker } from './utils/queue.js';
 import { updateEnrollmentsWorker } from './jobs/enrollmentsUpdate.js';
 import { userEnrollmentsUpdateWorker } from './jobs/userEnrollmentsUpdate.js';
 import { updateTeachersWorker } from './jobs/teacherUpdate.js';
+import { syncWorker } from './jobs/syncMatriculas.js';
 
 export const emailWorker = createWorker('Send:Email', sendEmailWorker);
 export const enrollmentsWorker = createWorker(
@@ -18,6 +19,7 @@ export const teachersUpdateWorker = createWorker(
   'Teacher:Update',
   updateTeachersWorker,
 );
+export const syncMatriculasWorker = createWorker('Matriculas:Sync', syncWorker);
 
 emailWorker.on('completed', (job) => {
   logger.info({
@@ -44,6 +46,14 @@ userEnrollmentsWorker.on('completed', (job) => {
 });
 
 teachersUpdateWorker.on('completed', (job) => {
+  logger.info({
+    msg: `[QUEUE] Job ${job.queueName} completed`,
+    id: job.id,
+    data: job.data,
+  });
+});
+
+syncMatriculasWorker.on('completed', (job) => {
   logger.info({
     msg: `[QUEUE] Job ${job.queueName} completed`,
     id: job.id,
